Ændrede int til double

+++ src/com/company/Database.java
@@ -5,8 +5,9 @@ public class Database {
 
     private ArrayList<Account> database = new ArrayList<Account>();
 
-    public void add(String ownerName, String accountNumber, int balance){
+    public void add(String ownerName, String accountNumber,double balance){
         Account account = new Account(ownerName, accountNumber, balance);
         database.add(account);
     }
+
 }
